Add some mod compatibility patches
This commit adds some compatibility patches for the mods ChatterReborn
and PacksHelper.

The PacksHelper patch should cover most things, however the
ChatterReborn patch only fixes the error spam with ChatterReborn. The
ChatterReborn patch will need to be improved in the future, as there
are other hardcoded things that need to be fixed.